Changed Position attributes from int to float to increase precision with decimal velocities. Began tuning running and jumping physics.

+++ saltar-core/src/main/java/com/thrashplay/saltar/component/FilledBlockRenderer.java
@@ -14,6 +14,6 @@ public class FilledBlockRenderer implements RenderableComponent {
     @Override
     public void render(Graphics graphics, GameObject gameObject) {
         Position position = gameObject.getComponent(Position.class);
-        graphics.fillCircle(position.getX(), position.getY(), 50, 0xffff0000);
+        graphics.fillCircle((int) position.getX(), (int) position.getY(), 50, 0xffff0000);
     }
 }
